+++ include/vec/order.h
@@ -2,8 +2,8 @@
 #define VEC_ORDER_H_IS_INCLUDED
 
 #include "vec/types.h"
-#include "vec/vec.h"
 
+VecNamespace vec_swap (Vec v, long a, long b);
 VecNamespace vec_reverse (Vec v);
 VecNamespace vec_sort (Vec v);